.requirements.txt
alembic==1.16.5
bcrypt==4.0.1
blinker==1.9.0
boto3==1.40.38
botocore==1.40.38
click==8.3.0
dnspython==2.8.0
email-validator==2.1.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.14.0
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.0.1
